requirements.txt
crewai
python-dotenv
requests
docling
fastapi[standard]
langchain_google_genai
google-generativeai